(from AES) Add (failing) tests
There are 3 toggles for known bugs

    bug_single_insecure_action = False # Do all Hosts have to have the same insecure.action?
    bug_forced_star = False            # Do we erroneously send replies in cleartext instead of TLS for unknown hosts?
    bug_404_routes = False             # Do we erroneously send 404 responses directly instead of redirect-to-tls first?

These toggles are all set to "False" in this commit; they exist to
make it possible for these tests to be used against 1.6.2/1.7.4-rc.1
where those bugs have "historical privilege" and don't count as
regressions.

The results of the added tests as of this commit:

  | HostCRDRootRedirectCongratulations  | pass                 |
  | HostCRDRootRedirectSlashMapping     | fail                 |
  | HostCRDRootRedirectRE2Mapping       | fail                 |
  | HostCRDRootRedirectECMARegexMapping | fail                 |
  | HostCRDForcedStar                   | pass                 |
  | TLSInheritFromModule                | doesn't become ready |

None of the errors have to do with the `bug_*` toggles; setting any of
the toggles to "True" wouldn't make a difference.

This should cover the issues:
  https://github.com/datawire/ambassador/issues/2957
  https://github.com/datawire/ambassador/issues/2996

I didn't write automated tests for
https://github.com/datawire/ambassador/issues/2984 since any test I
can think of writing that might pass for 1.6.2 and 1.7.4-rc.1 is just
testing for buggy behavior.  I think we need to have a conversation
about what the correct behavior is.